Add a 'manualFix' custom event
Implements https://github.com/jonathanpeppers/inclusive-code-comments/issues/24

I implemented the bulk of the logic for this in
`src-packed/validator.js`. A `hasFoundSuggestionsBefore` variable
tracks the state -- and we report `manualFix` when a fix is resolved
after a previous one was found. Doing it this way, I was able to write
tests and test the interaction between the `appliedSuggestion` event.

The only problem I found was the way `ltAssistant.js` sent the
`appliedSuggestion` event from *inside* web pages, instead of from the
extension in the background. So we ended up with two instances of
`src-packed/validator.js` and we get two events: `appliedSuggestion`
then `manualFix`!

To solve this, I use `sendMessage()` to send from the page:

    browser.runtime.sendMessage({ command: "APPLIED_SUGGESTION", appliedSuggestions: appliedSuggestions });

Then in the background `extension-main.js` has a
`_onAppliedSuggestionMessage()` method to receive the message.

So the background extension sends both messages and can properly
handle the state now.

Other changes:

* `appinsights` can now be loaded in tests since 9ee7a949, so I used a
  `loadAppInsights()` method. We don't need null checks anymore for it.