[appinsights] send all data only from the background extension (#38)
This uses `browser.runtime.sendMessage()` to get a message from inside the page to the background extension.

So we can pass things like:

    browser.runtime.sendMessage({ command: "TRACK_PAGE_VIEW", location: { name: document.title, uri: window.location.href } });

Then in the background extension we pass to Application Insight. We have a stable user id this way.

NOTE: we still don't record full URLs, only Github orgs & repositories.