Add custom negative sentiments to suggestions.js (#59)
Context: #46

The Text Analysis service does not seem to pick up some regional
phrases which are negative. To work around this for now lets start adding
additional negative sentiments to the suggestions.js.

We do not need to provide replacement suggestions, we can leave them empty
and the extension will flag them as negative as usual.