[telemetry] better logic for the 'manualFix' event (#61)
After looking at #57, the number of `manualFix` events seemed low.

It looked like we only reported this event when there was an error in
the past, and now all errors are fixed. Change this value to an
integer of the number of errors, and we can report `manualFix` if the
number of errors *goes down*.

This way we'll get the `manualFix` event if you have more than one
error and fix at least one.