Add ignore list for certain negative phrases
Partially fixes: https://github.com/jonathanpeppers/inclusive-code-comments/issues/47

Adds support for ignoring certain words or phrases with a negative
sentiment.  These ignorable words will be replaced with an empty string
of equal length in the comment before sending it to the Text Analytics
service.

Only a couple of phrases have been added to this list, and it can be
expanded in the future as needed.