Add suggestions for "kill", and "stuck" (#104)
Fixes: https://github.com/jonathanpeppers/inclusive-code-comments/issues/82

Issue #82 appears to be Azure Text Analytics sentiment analysis, but
as a way to address this. We should recommend using "stop" or "cancel"
in favor of "kill".

I tested "killing it", and it appears things are working as intended
when one problematic phrase includes another inside of it. "kill" is a
substring of "killing it", but it appears to work fine.