Suggest "can probably" for "should just" (#171)
This PR is adding a category of common wording pattern that can be softened.

"should just" is a very common saying in my native language. I get feedback from native English speaker that it can be soften by changing to "can probably". e.g. "We should just delete this section" -> "We can probably delete this section".