Project clean up - Part 1 (#157)
This PR aims to address #111. But the code base is now tightly integrated with the original LanguageTool, so it's less likely to be done in one PR; Instead, we will clean it up piece by piece. This is the first part that cleans those easy targets from the code base.

* Remove unused l10n strings

Co-authored-by: AJ Hsu <ajhsu@microsoft.com>
Co-authored-by: Jonathan Peppers <jonathan.peppers@microsoft.com>