Remove more `window.setInterval()` calls (#191)
While working on PR #90, I'm trying to remove as many
`window.setInterval()` calls as we can -- that aren't really
necessary. This should help get us on Chrome V3, general performance,
remove weird code, etc.

There appears to be a timer for checking the "health" of the
extension, and then removing various UI elements from the page.
Nothing appeared to break when removing it? Let's go with it.

There was a second timer for setting up various hacks for a thing
called TinyMCE:

https://www.tiny.cloud/

This appears to be some WYSIWYG editor that websites use? I removed
this code, as it appears that GitHub and Azure DevOps still work fine
without these hacks in place.