Use `chrome.runtime.connect` for messaging
Context: https://github.com/jonathanpeppers/HelloChrome

In a "hello world" extension, I was able to find some Chrome APIs
where:

* service worker falls asleep
* content script from page sends a message
* service worker wakes up
* everything works!

I also found a nice page to debug this behavior:

edge://serviceworker-internals/?devtools

This page will list all the service workers, and there is a `Stop`
button to put them to sleep.

The key here is instead of:

    //content script
    var response = await chrome.runtime.sendMessage({greeting: "hello"});

    //background.js
    chrome.runtime.onConnect.addListener((message, sender, sendResponse) => {
        sendResponse({greeting: "goodbye"});
    });

Do this:

    // content script
    var port = chrome.runtime.connect({name: "test"});
    port.postMessage({greeting: "hello"});

    // background.js
    chrome.runtime.onConnect.addListener(function(port) {
        port.onMessage.addListener(function(message) {
            port.postMessage({greeting: "goodbye"});
        });
    });

The big problem here now is that the messages are asynchronous.

This is only about 1/2 working. I need to somehow collapse the callbacks
into a promise (ala `TaskCompletionsSource` in .NET). I'll need to learn
the JavaScript on how to do that.

Other changes:

* removed "offscreen" page as it is no longer needed