Allow V3 extension to "wake up" appropriately
Context: https://github.com/jonathanpeppers/inclusive-code-reviews-browser/pull/285

This is a smaller patch that also seems to fix the general problem here
that exists for the Chrome V3 extension (on Chrome store):

1. Extension is working.

2. At some point Chrome "stops" the background service worker.

3. Extension no longer works until you click the "Restart" button in the
   edge://extensions page.

One problem here was we loaded all the scripts in `self.oninstall`,
where we could just load them directly instead. This was basically only
making the extension work for a period of time after it was initially
installed.

You can test this behavior with the "Stop" button on this page:

    edge://serviceworker-internals/?devtools