Disable clicking the second popup
We generally click on the underlined word in text, but there is a
"second popup" that is not fully functional and a bit odd:

* It shows negative sentiment with text underlined, confusing.
* It has options to change language?!?
* It has options to dock the popup all over the web page.

There are also a couple issues opened, related:

* https://github.com/jonathanpeppers/inclusive-code-reviews-browser/issues/338
* https://github.com/jonathanpeppers/inclusive-code-reviews-browser/pull/329

For now, let's just disable the second popup? I also disabled the
hover effect for the pointer and updated the tooltip text for this
button.