[telemetry] add `ai.user.id` to identify unique installs
Fixes: https://github.com/jonathanpeppers/inclusive-code-reviews-browser/issues/405

Store an anonymous `self.crypto.randomUUID()` value, we can use to
just find out "unique installs" of the browser extension.

This used to work automatically by AppInsights, but it either broke
when we moved to Chrome V3 or a dependabot update.

With this change I see in the network tab:

    ai.user.id : "ab57ffa4-f33f-46ee-b4fc-94da8fb6c5c5"

I see the same number throughout the session, and it's different
if I uninstall and reinstall the extension.